// Projects/Tamin/Eservices-UI/src/app/models/pensioner/edictPaymentDetail.model.ts
export class EdictPaymentDetailModel {
  id: string;
  clpType: string;
  tprDesc: string;
  sumAmount: number;
  infDesc: any;
  textNumber: string;
// edictCalculationHead: any;
// paymentItemType: any;
// paymentAmount: string;
// startDate: string;
// monthCount: string;
// tafsiliCode: string;
// counts: string;
// bedOrBes: string;
// difference: string;
// postponeStartDate: string;
// postponeEndDate: string;
// insuranceNumber: string;
// branch: any;
// totalCode: string;
// moeenCode: string;
// activityCode: string;
}